split bit, dec, inc
