+++ src/mnemonic.cpp
@@ -1,7 +1,10 @@
 #include "common.h"
 #include "mnemonic.h"
+#include "mnemonic_bit.hpp"
 #include "mnemonic_branch.hpp"
 #include "mnemonic_calc.hpp"
+#include "mnemonic_dec.hpp"
+#include "mnemonic_inc.hpp"
 #include "mnemonic_load.hpp"
 #include "mnemonic_other.hpp"
 #include "mnemonic_shift.hpp"
